+++ package.json
@@ -4,7 +4,7 @@
   "description": "Implements the Jaccard algorithm for finding the similarity coefficient between sentences, by Marcel",
   "main": "app/index.js",
   "scripts": {
-    "test": "./node_modules/.bin/mocha --reporter spec",
+    "test": "./node_modules/.bin/mocha 'tests/*.test.js' --reporter spec",
     "start": "node ./app/index.js"
   },
   "repository": {
